Restructured users directory. It now has different directories for different request methods
