+++ src/endpoints/user/__init__.py
@@ -1,7 +1,7 @@
-from .delete_current_user import delete_current_user
-from .delete_user_id import delete_user_by_id
-from .filter_user import filter_user
-from .get_users import get_user_by_id
+from .delete.delete_current_user import delete_current_user
+from .delete.delete_user_id import delete_user_by_id
+from .get.filter_user import filter_user
+from .get.get_users import get_user_by_id
+from .put.update_current_user import update_current_user
+from .put.update_user_id import update_current_user_by_id
 from .router_init import router
-from .update_current_user import update_current_user
-from .update_user_id import update_current_user_by_id
