+++ src/endpoints/user/delete/delete_current_user.py
@@ -5,10 +5,10 @@ from sqlalchemy import delete
 from sqlalchemy.orm import Session
 from starlette import status
 
-from ...dependencies import get_current_user, get_db
-from ...models.user import User
-from .exceptions import db_not_available
-from .router_init import router
+from ....dependencies import get_current_user, get_db
+from ....models.user import User
+from ..exceptions import db_not_available
+from ..router_init import router
 
 
 @router.delete("/", status_code=status.HTTP_204_NO_CONTENT)
