+++ src/endpoints/user/put/update_current_user.py
@@ -5,12 +5,12 @@ from sqlalchemy import select
 from sqlalchemy.orm import Session
 from starlette import status
 
-from ...dependencies import get_current_user, get_db, verify_password
-from ...models.user import User
-from ...schemas.update_user import UpdateUserSchema
-from .exceptions import invalid_data, old_pass_not_matched
-from .router_init import router
-from .user_utils import update_user
+from ....dependencies import get_current_user, get_db, verify_password
+from ....models.user import User
+from ....schemas.update_user import UpdateUserSchema
+from ..exceptions import invalid_data, old_pass_not_matched
+from ..router_init import router
+from ..user_utils import update_user
 
 
 @router.put("/", status_code=status.HTTP_200_OK, response_model=None)
